+++ guillotina/api/files.py
@@ -10,7 +10,6 @@ from guillotina.interfaces import IFileManager
 from guillotina.interfaces import IResource
 from guillotina.interfaces import IStaticFile
 
-import aiohttp
 import mimetypes
 
 
@@ -22,9 +21,9 @@ class DefaultGET(DownloadService):
             filepath = str(self.context.file_path.absolute())
             filename = self.context.file_path.name
             with open(filepath, 'rb') as f:
-                resp = StreamResponse(headers=aiohttp.MultiDict({
+                resp = StreamResponse(headers={
                     'CONTENT-DISPOSITION': 'attachment; filename="%s"' % filename
-                }))
+                })
                 resp.content_type = mimetypes.guess_type(filename)
                 data = f.read()
                 resp.content_length = len(data)